compaction_manager: use double for backlog everywhere
Avi says:
"The backlog is a large number that changes slowly, so float
might not have enough resolution to track small changes.

For example, if the backlog is 800GB and changes less than 100kB, then
we won't see a change (float resolution is 2^23 ~ 1:8,000,000).

This is outside the normal range of values (usually the backlog changes
a lot more than 100kB per 15-second period), so it will work, but better
to be more careful."

Message-Id: <20200615150621.17543-1-raphaelsc@scylladb.com>

+++ sstables/compaction_manager.hh
@@ -87,7 +87,7 @@ private:
 
     stats _stats;
     seastar::metrics::metric_groups _metrics;
-    float _last_backlog = 0.0f;
+    double _last_backlog = 0.0f;
 
     std::list<lw_shared_ptr<sstables::compaction_info>> _compactions;
 
@@ -258,7 +258,7 @@ public:
     // called before compaction seals sstable and such and after all compaction work is done.
     void on_compaction_complete(compaction_weight_registration& weight_registration);
 
-    float backlog() {
+    double backlog() {
         return _backlog_manager.backlog();
     }
